==> 01 players.ipynb <==
# Wyscout Data for Inter 2017-2018

In questo notebook vengono selezionate le informazioni relativi ai calciatori dell'Inter nella stagione 2017-2018

## utilities

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import util
path = "C:/Users/Mauro/OneDrive/Documenti/Football/playerank-master"
inter_id = 3161

### players

**GET ALL THE PLAYERS**

players = pd.read_json(os.path.join(path, 'data/players.json'))
print(len(players))
players.head(3)

**GET ALL THE INTER PLAYERS**

inter_players = players[players.currentTeamId == inter_id]
inter_players.head(3)

**NORMALIZE COLUMNS WHICH ARE NESTED JSON**

inter_players = util.normalize_column(inter_players, ["birthArea", "passportArea", "role"])
inter_players.head(3)

**DECODE NAME STRING OF THE PLAYERS**

inter_players = util.decode_string(inter_players, ["firstName", "lastName", "middleName", "shortName"])
inter_players.head(3)

**EXPLORE INTER PLAYERS**

print("There are", inter_players.shape[0], "players")

print("For each player we have these informations")
inter_players.dtypes

inter_players = inter_players[['wyId', 'role_name', 'shortName']].sort_values(by=["role_name", "shortName"])
inter_players

**EXPORT USEFUL DATA IN CSV FILE FOR NEXT ANALYSIS**

inter_players.to_csv("inter_players.csv", index_label=False)
