==> 02 team.ipynb <==
# Wyscout Data for Inter 2017-2018

In questo notebook vengono selezionate le informazioni relative alla squadra dell'Inter

## utilities

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from pandas.io.json import json_normalize
import util
path = "C:/Users/Mauro/OneDrive/Documenti/Football/playerank-master"
inter_id = 3161

### teams

**GET ALL THE TEAMS**

teams = pd.read_json(os.path.join(path, 'data/teams.json'))
len(teams)
teams.head(3)

**GET INFO ABOUT INTER TEAM**

inter_team = teams[teams.wyId == inter_id]
inter_team

**NORMALIZE COLUMNS THAT ARE NESTED JSON**

inter_team = util.normalize_column(inter_team, ["area"])
inter_team

**EXPLORE INFORMATIONS**

print("For each team we have these informations")
inter_team.dtypes

inter_team[['wyId', 'name']]
