==> 03 matches.ipynb <==
# Wyscout Data for Inter 2017-2018

In questo notebook vengono selezionate le informazioni relative ai match disputati dall'Inter nella stagione 2017-2018

## utilities

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from pandas.io.json import json_normalize
import util
path = "C:/Users/Mauro/OneDrive/Documenti/Football/playerank-master"
inter_id = 3161

### matches

**GET ALL SERIE A MATCHES**

matches = pd.read_json(os.path.join(path, 'data/matches/matches_Italy.json'))
len(matches)
matches.head(3)

**GET ALL INTER MATCHES**

inter_matches = matches[matches.label.str.contains("Internazionale")]
inter_matches.head(3)

**NORMALIZE REFEREES DATA**

inter_matches = util.get_referees(inter_matches, "referees")
inter_matches.head(3)

**GET TEAMS DATA FOR EVERY INTER MATCH IN A DICTIONARY**

teamsData = util.get_teamsData(inter_matches)
match_id = inter_matches[inter_matches.label.str.contains("Lazio - Internazionale")].wyId[0]
teamsData[match_id]

**GET ALL THE FORMATIONS ABOUT INTER MATCHES (OPPONENT TOO)**

formationsData = util.get_formations(teamsData)
match_id = inter_matches[inter_matches.label.str.contains(" - Milan")].wyId.iloc[0]
formationsData[formationsData.matchId == match_id]

**EXPLORE INFORMATION ABOUT INTER MATCHES**

print("There are", inter_matches.shape[0], "matches")

print("For each match we have these informations")
inter_matches.dtypes

inter_matches = inter_matches[['wyId', 'gameweek', 'label']].sort_values(by="gameweek")
inter_matches

path = "C:/Users/Mauro/OneDrive/Documenti/Football/Inter_2017_2018"
inter_players = pd.read_csv(os.path.join(path, "inter_players.csv"))
derby_lineups = formationsData.query("matchId == 2576034 and team == 'home'")
pd.merge(derby_lineups, inter_players, how="left", left_on="playerId", right_on="wyId")[['playerId', 'shortName']]

**EXPORT USEFUL DATA IN CSV FILE**

inter_matches.to_csv("inter_matches.csv", index_label=False)
formationsData.to_csv("inter_matches_formations.csv", index_label=False)
