==> 04 events.ipynb <==
# Wyscout Data for Inter 2017-2018

In questo notebook vengono selezionate le informazioni relative agli eventi dei match disputati dall'Inter nella stagione 2017-2018

## utilities

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from pandas.io.json import json_normalize
import util
path = "C:/Users/Mauro/OneDrive/Documenti/Football/playerank-master"
inter_id = 3161

### events

**GET ALL THE SERIE A MATCH EVENTS** 

events = pd.read_json(os.path.join(path, 'data/events/events_Italy.json'))
len(events)
events.head(3)

**GET ALL THE INTER MATCH EVENTS**

path = "C:/Users/Mauro/OneDrive/Documenti/Football/Inter_2017_2018"
inter_matches = pd.read_csv(os.path.join(path, "inter_matches.csv"))
inter_matches_events = events.query("matchId in " + str(list(inter_matches.wyId)))
inter_matches_events.head(3)

**NORMALIZE EVENTS LOCATION**

inter_matches_events = util.get_event_locations(inter_matches_events, "positions")
inter_matches_events.head(3)

**CREATE BOOLEAN COLUMNS FOR EVENTS TAGS**

inter_matches_events = util.get_event_tags(inter_matches_events, "tags")
inter_matches_events.head(3)

**GET GOALS ABOUT DERBY AT HOME**

match_id = inter_matches[inter_matches.label.str.contains(" - Milan")].wyId.iloc[0]
derby_goals = inter_matches_events.query("matchId == " + str(match_id) + " and ((teamId == " + str(inter_id) + " and goal == 'yes' and accurate == 'yes') or (teamId != " + str(inter_id) + " and own_goal == 'yes'))")
inter_players = pd.read_csv(os.path.join(path, "inter_players.csv"))
pd.merge(derby_goals, inter_players, how="left", left_on="playerId", right_on="wyId")[['playerId', 'shortName']]

**EXPLORE DATA ABOUT INTER MATCH EVENTS**

match_id = inter_matches[inter_matches.label.str.contains(" - Juventus")].wyId.iloc[0]
int_juv_events = inter_matches_events.query("matchId == " + str(match_id))
home_events = int_juv_events.query("teamId == " + str(inter_id))
away_events = int_juv_events.query("teamId != " + str(inter_id))
print("In Inter-Juventus there are", home_events.shape[0], "events associated to Inter")
print("In Inter-Juventus there are", away_events.shape[0], "events associated to Juventus")

print("For every event we have these informations")
inter_matches_events.dtypes

**EXPORT USEFUL DATA FOR NEXT ANALYSIS IN CSV FILE**

inter_matches_events.to_csv(os.path.join(path, "inter_matches_events.csv"), index_label=False)
