==> 07 shot_analysis.ipynb <==
# Wyscout Data for Inter 2017-2018

In questo notebook vengono analizzati i tiri effettuati dai calciatori dell'Inter durante la stagione 2017-2018 <br>
In particolare: <br>
1) conteggio dei tiri per ogni calciatore durante la stagione <br>
2) conteggio dei goal per ogni calciatore durante la stagione <br>
3) rank dei calciatore per percentuale realizzativa o numero di tiri per goal 

## utilities

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import util
import warnings
warnings.simplefilter("ignore")
path = "C:/Users/Mauro/OneDrive/Documenti/Football/Inter_2017_2018"
inter_id = 3161

**COUNT OF TOTAL SHOTS FOR EACH PLAYER IN EACH MATCH**

inter_matches_events = pd.read_csv(os.path.join(path, "inter_matches_events.csv"))
inter_matches_events.head(3)

#filter by Inter shots (eventId == 10 and teamId == 3161) 
shots = inter_matches_events.query("eventId == 10 and teamId == " + str(inter_id))
shots.head(3)

#merge with match data and player data
inter_matches = pd.read_csv(os.path.join(path, "inter_matches.csv"))
inter_players = pd.read_csv(os.path.join(path, "inter_players.csv"))
shots = pd.merge(shots, inter_matches[["wyId", "gameweek", "label"]], left_on="matchId", right_on="wyId")
shots = pd.merge(shots, inter_players[["wyId", "shortName", "role_name"]], left_on="playerId", right_on="wyId")
shots.head(3)

#group by for match (matchId) and player (playerId)
total_shots = shots.groupby(["matchId", "gameweek", "label", "playerId", "shortName"]).size().reset_index()
total_shots.columns = ["matchId", "gameweek", "label", "playerId", "shortName", "#shots"]
total_shots.head(3)

total_shots = pd.DataFrame(total_shots.groupby(["playerId", "shortName"])["#shots"].sum().sort_values(ascending=False))
total_shots

goals = shots.query("goal == 'yes' and accurate == 'yes'").groupby(["matchId", "gameweek", "label", "playerId", "shortName"]).size().reset_index()
goals.columns = ["matchId", "gameweek", "label", "playerId", "shortName", "#goals"]
goals = pd.DataFrame(goals.groupby(["playerId", "shortName"])["#goals"].sum().sort_values(ascending=False))
goals

**CALCULATE SHOT PER GOAL AND SHOTS TO GOAL PERCENTAGE (MIN 10 SHOTS)**

df = pd.merge(total_shots, goals, left_index=True, right_index=True)
df = df[df["#shots"] > 10]
df["shots_per_goal"] = round(df["#shots"] / df["#goals"], 2)
df["shots_to_goal_percentage"] = round(df["#goals"] / df["#shots"] * 100, 2)
df.sort_values(by="shots_to_goal_percentage", ascending=False)
